requirements/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pydantic==2.6.3
jdatetime==4.1.1
